#220 rename Python module
The previous module name was too vague. The file_io name provides a
better hint at the module's purpose.
